refactor(videos): move video URL utils from bunny.server to utils
- Moved getVideoEmbedUrl and getVideoThumbnailUrl functions from app/lib/bunny.server.ts to app/lib/utils.ts
- Updated imports in video-card.tsx and videos-list.tsx to import these utils from utils instead of bunny.server
- Removed redundant code from bunny.server.ts

This refactor centralizes video URL utility functions in utils.ts for better code organization and maintainability.
